Refactor test compute
